+++ misc/multiaddr/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.9.5 [2020-11-14]
+
+- Limit initial memory allocation in `visit_seq`.
+  [PR 1833](https://github.com/libp2p/rust-libp2p/pull/1833).
+
+# 0.9.4 [2020-11-09]
+
+- Update dependencies.
+
 # 0.9.3 [2020-10-16]
 
 - Update dependencies.